Move `common` back to `bn256`
