+++ src/bn256/common.rs
